requirements.txt
Django==1.10.3
django-extensions==1.7.4
psycopg2==2.6.2
six==1.10.0
django-debug-toolbar==1.6
python-social-auth==0.2.21